requirements.txt
asgiref==3.3.0
beautifulsoup4==4.9.3
config==0.3.9
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.17
django-bootstrap3==12.0.0
django-crispy-forms==1.7.0
django-heroku==0.3.1
django-registration==2.4.1
djangorestframework==3.8.2
gunicorn==20.0.4
Pillow==8.0.1
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.4
soupsieve==2.0.1
sqlparse==0.4.1
whitenoise==5.2.0
